Move DbType and DbSettings classes into configuration subpackage
