+++ rabbit-core/src/main/java/org/ohdsi/databases/DBConnectorInterface.java
@@ -2,6 +2,8 @@ package org.ohdsi.databases;
 
 import org.ohdsi.databases.configuration.ConfigurationField;
 import org.ohdsi.databases.configuration.DBConfiguration;
+import org.ohdsi.databases.configuration.DbSettings;
+import org.ohdsi.databases.configuration.DbType;
 import org.ohdsi.utilities.files.IniFile;
 import org.ohdsi.utilities.files.Row;
