+++ whiterabbit/src/main/java/org/ohdsi/whiterabbit/scan/SourceDataScan.java
@@ -36,7 +36,8 @@ import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.commons.io.FileUtils;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import org.ohdsi.databases.DbType;
+import org.ohdsi.databases.configuration.DbSettings;
+import org.ohdsi.databases.configuration.DbType;
 import org.ohdsi.databases.RichConnection;
 import org.ohdsi.databases.QueryResult;
 import org.ohdsi.databases.*;
