+++ whiterabbit/src/test/java/org/ohdsi/whiterabbit/scan/SourceDataScanPostgreSQLGuiIT.java
@@ -10,7 +10,7 @@ import org.assertj.swing.fixture.FrameFixture;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
-import org.ohdsi.databases.DbType;
+import org.ohdsi.databases.configuration.DbType;
 import org.ohdsi.whiterabbit.Console;
 import org.ohdsi.whiterabbit.WhiteRabbitMain;
 import org.ohdsi.whiterabbit.gui.LocationsPanel;
@@ -26,7 +26,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.ohdsi.databases.DbType.POSTGRESQL;
+import static org.ohdsi.databases.configuration.DbType.POSTGRESQL;
 import static org.ohdsi.whiterabbit.scan.SourceDataScanPostgreSQLIT.createPostgreSQLContainer;
 
 @ExtendWith(GUITestExtension.class)
