+++ whiterabbit/src/test/java/org/ohdsi/whiterabbit/scan/SourceDataScanSnowflakeGuiIT.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.ohdsi.databases.SnowflakeConnector;
 import org.ohdsi.databases.SnowflakeTestUtils;
-import org.ohdsi.databases.DbType;
+import org.ohdsi.databases.configuration.DbType;
 import org.ohdsi.whiterabbit.Console;
 import org.ohdsi.whiterabbit.WhiteRabbitMain;
 import org.testcontainers.containers.GenericContainer;
@@ -30,7 +30,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.ohdsi.databases.DbType.SNOWFLAKE;
+import static org.ohdsi.databases.configuration.DbType.SNOWFLAKE;
 import static org.ohdsi.whiterabbit.scan.SourceDataScanSnowflakeIT.*;
 
 @ExtendWith(GUITestExtension.class)
